+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CO
 
 ### Fixed
 
+* Json formatter now works with tagged examples
+  ([#1621](https://github.com/cucumber/cucumber-js/issues/1621)
+   [#1651](https://github.com/cucumber/cucumber-js/pull/1651))
+
 ## [7.2.1] (2021-04-21)
 
 ### Fixed